+++ EZ.Apple/Extensions/Extension.swift
@@ -1,0 +1,20 @@
+//
+//  Extension.swift
+//  EZ.Apple
+//
+//  Created by seunghoria on 2023/05/07.
+//
+
+import SwiftUI
+
+struct Extension: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct Extension_Previews: PreviewProvider {
+    static var previews: some View {
+        Extension()
+    }
+}
