+++ EZ.Apple/ShareView/matchedGeometryEffect.swift
@@ -1,0 +1,21 @@
+//
+//  matchedGeometryEffect.swift
+//  EZ.Apple
+//
+//  Created by seunghoria on 2023/05/07.
+//
+
+import SwiftUI
+
+struct matchedGeometryEffect: View {
+    var body: some View {
+        Text("Hello, Seez!")
+            .frame(maxWidth: .infinity, alignment: .leading)
+    }
+}
+
+struct matchedGeometryEffect_Previews: PreviewProvider {
+    static var previews: some View {
+        matchedGeometryEffect()
+    }
+}
